+++ changelog.md
@@ -7,6 +7,7 @@ Summary of implemented features classified by release (github tag)
 - fix issue #44 : sort posts by date + tests
 - fix issue #24 : sort posts with keyword in title + tests
 - implement limit & offset on route GET `/api/v1/posts` + tests
+- fix issue #84 : prevent creation of post if empty string in title or content
 
 ## [0.3.0] - 2020-04-30